refactor: Relocate components and theme to designsystem module
- Move composables (`DiveSoptButton`, `DiveSoptTextField`, `FormField`) from `ui.core.component` to `core.designsystem.component`.
- Move theme-related files (`Color.kt`, `Type.kt`, `Theme.kt`) from `ui.theme` to `core.designsystem.theme`.
- Update import paths in `MainActivity`, `SignInActivity`, and `SignUpActivity` to reflect the new locations.
