+++ app/src/main/java/com/sopt/dive/core/designsystem/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.sopt.dive.ui.theme
+package com.sopt.dive.core.designsystem.theme
 
 import androidx.compose.ui.graphics.Color
