+++ app/src/main/java/com/sopt/dive/core/designsystem/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.sopt.dive.ui.theme
+package com.sopt.dive.core.designsystem.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
