+++ app/src/main/java/com/sopt/dive/core/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.sopt.dive.ui.theme
+package com.sopt.dive.core.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
